refactor: add loop class
