+++ packages/client/src/game/services/Game.ts
@@ -1,17 +1,9 @@
-import { Entity, Projectile, Tank } from '../entities';
-import {
-  Direction,
-  GameSettings,
-  LoopDelays,
-  LoopIntervals,
-  MainMenuState,
-  ScenarioEvent,
-  ScreenType,
-} from '../typings';
+import { Entity } from '../entities';
+import { Direction, GameSettings, MainMenuState, ScenarioEvent, ScreenType } from '../typings';
 import { Overlay } from '../ui';
 import { levels } from './../data/levels';
-import { ControllerEvent, EntityEvent } from './../typings/index';
-import { Controller, resources, Scenario, View, Zone } from './';
+import { ControllerEvent } from './../typings/index';
+import { Controller, Loop, resources, Scenario, View, Zone } from './';
 import { AudioManager } from './AudioManager';
 import { KeyBindingsArrows, KeyBindingsWasd } from './KeyBindings';
 
@@ -19,20 +11,15 @@ export class Game {
   static __instance: Game;
   inited = false;
   paused = false;
-  zone!: Zone;
-  view!: View;
-  audioManager: AudioManager = new AudioManager();
-  overlay!: Overlay;
+  loop: Loop;
+  zone: Zone;
+  view: View;
+  audioManager: AudioManager;
+  overlay: Overlay;
   scenario: Scenario | undefined;
-  controllerAll!: Controller;
-  controllerWasd!: Controller;
-  controllerArrows!: Controller;
-  loopProcess: ReturnType<typeof setTimeout> | null = null;
-  loopTimeMs = 25;
-  loopCount = 0;
-  loopDelays: LoopDelays = {};
-  loopIntervals: LoopIntervals = {};
-  loopEntities: Set<Tank | Projectile> = new Set();
+  controllerAll: Controller;
+  controllerWasd: Controller;
+  controllerArrows: Controller;
   settings: GameSettings = { width: 56, height: 56, boundarySize: 2 };
   screen: ScreenType = ScreenType.LOADING;
   mainMenuState = MainMenuState.SINGLEPLAYER;
@@ -40,9 +27,11 @@ export class Game {
   maxLevels = levels.length;
 
   private constructor() {
+    this.loop = new Loop();
     this.zone = new Zone(this.settings);
     this.view = new View(this.settings);
     this.overlay = new Overlay(this);
+    this.audioManager = new AudioManager();
     this.controllerAll = new Controller({ ...KeyBindingsWasd, ...KeyBindingsArrows });
     this.controllerWasd = new Controller(KeyBindingsWasd);
     this.controllerArrows = new Controller(KeyBindingsArrows);
@@ -65,7 +54,7 @@ export class Game {
 
   load(root: HTMLElement | null) {
     this.createView(root);
-    this.startLoop();
+    this.loop.load();
     this.controllerAll.load();
     this.controllerWasd.load();
     this.controllerArrows.load();
@@ -73,9 +62,7 @@ export class Game {
   }
 
   unload() {
-    this.clearLoopEntities();
-    this.clearLoopDelays();
-    this.stopLoop();
+    this.loop.unload();
     this.controllerAll.unload();
     this.controllerWasd.unload();
     this.controllerArrows.unload();
@@ -86,9 +73,7 @@ export class Game {
     if (this.scenario) {
       delete this.scenario;
     }
-    this.clearLoopEntities();
-    this.loopIntervals = {};
-    this.clearLoopDelays();
+    this.loop.reset();
     this.view.reset();
     this.zone.reset();
     this.audioManager.reset();
@@ -102,108 +87,10 @@ export class Game {
   }
 
   addEntity(entity: Entity) {
+    this.loop.add(entity);
     this.view.add(entity);
     this.zone.add(entity);
     this.audioManager.add(entity);
-    this.registerTimerHandlers(entity);
-    if (entity instanceof Tank) {
-      this.loopEntities.add(entity);
-    } else if (entity instanceof Projectile) {
-      const tempLoopEntitiesArray = Array.from(this.loopEntities);
-      tempLoopEntitiesArray.unshift(entity);
-      this.loopEntities = new Set(tempLoopEntitiesArray);
-    }
-  }
-
-  convertTimeToLoops(delay: number) {
-    return Math.floor(delay / this.loopTimeMs);
-  }
-
-  /** Аналог setTimeout, который работает через игровой цикл. */
-  setLoopDelay(callback: () => void, delay: number) {
-    const loopMark = this.loopCount + this.convertTimeToLoops(delay);
-
-    if (!this.loopDelays[loopMark]) {
-      this.loopDelays[loopMark] = [];
-    }
-
-    this.loopDelays[loopMark].push(callback);
-  }
-
-  clearLoopDelays() {
-    this.loopCount = 0;
-    this.loopDelays = {};
-  }
-
-  /** Аналог setInterval, который работает через игровой цикл. */
-  setLoopInterval(callback: () => void, delay: number, intervalName: string) {
-    this.loopIntervals[intervalName] = {
-      loopCounter: 0,
-      targetLoop: this.convertTimeToLoops(delay),
-      callback: callback,
-    };
-
-    return intervalName;
-  }
-
-  clearLoopInterval(intervalName: string) {
-    if (intervalName in this.loopIntervals) {
-      delete this.loopIntervals[intervalName];
-    }
-  }
-
-  registerTimerHandlers(entity: Entity) {
-    entity.on(EntityEvent.SET_LOOP_DELAY, this.setLoopDelay.bind(this));
-    entity.on(EntityEvent.SET_LOOP_INTERVAL, this.setLoopInterval.bind(this));
-    entity.on(EntityEvent.CLEAR_LOOP_INTERVAL, this.clearLoopInterval.bind(this));
-  }
-
-  checkLoopDelays() {
-    if (this.loopDelays[this.loopCount]) {
-      const delayedCallbacks = this.loopDelays[this.loopCount];
-      for (const callback of delayedCallbacks) {
-        callback();
-      }
-      delete this.loopDelays[this.loopCount];
-    }
-  }
-
-  checkLoopIntervals() {
-    Object.values(this.loopIntervals).forEach(interval => {
-      if (interval.loopCounter === interval.targetLoop) {
-        interval.callback();
-        interval.loopCounter = 0;
-        return;
-      }
-      interval.loopCounter++;
-    });
-  }
-
-  clearLoopEntities() {
-    for (const entity of this.loopEntities) {
-      entity.despawn();
-    }
-    this.loopEntities = new Set();
-  }
-
-  loop() {
-    const cycleStartTime = performance.now();
-    let nextCycleDelay = this.loopTimeMs;
-    ++this.loopCount;
-    this.checkLoopDelays();
-    this.checkLoopIntervals();
-    for (const entity of this.loopEntities) {
-      entity.update();
-      if (entity.shouldBeDestroyed) {
-        this.loopEntities.delete(entity);
-      }
-    }
-    nextCycleDelay -= cycleStartTime - performance.now();
-    if (nextCycleDelay < 0) {
-      nextCycleDelay = 0;
-    }
-
-    this.loopProcess = setTimeout(this.loop.bind(this), nextCycleDelay);
   }
 
   togglePause(newState: boolean | null = null) {
@@ -213,29 +100,17 @@ export class Game {
 
     if (newState === false || this.paused) {
       this.overlay.clearScreen();
-      this.startLoop();
+      this.loop.start();
       this.controllerWasd.load();
       this.controllerArrows.load();
     } else if (newState === true || !this.paused) {
       this.overlay.show(ScreenType.PAUSE);
-      this.stopLoop();
+      this.loop.stop();
       this.controllerWasd.unload();
       this.controllerArrows.unload();
     }
     this.paused = !this.paused;
     this.audioManager.emit('pause', this.paused);
-  }
-
-  startLoop() {
-    this.stopLoop();
-    this.loop();
-  }
-
-  stopLoop() {
-    if (this.loopProcess) {
-      clearTimeout(this.loopProcess);
-      this.loopProcess = null;
-    }
   }
 
   initLoading() {
@@ -362,7 +237,7 @@ export class Game {
     this.overlay.show(this.screen, startAnimationDelay);
 
     /** Стартуем сценарий после окончания анимации */
-    this.setLoopDelay(() => {
+    this.loop.setLoopDelay(() => {
       /** Инициализируем инстанс сценария */
       this.scenario = new Scenario(this)
         .on(ScenarioEvent.GAME_OVER, () => {
