+++ packages/client/src/game/services/index.ts
@@ -1,5 +1,6 @@
 export { Controller } from './Controller';
 export { Game } from './Game';
+export { Loop } from './Loop';
 export { MapManager } from './MapManager';
 export { resources } from './Resources/Resources';
 export { Scenario } from './Scenario';
